update naming on front end
