+++ client/src/components/App.js
@@ -2,11 +2,11 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 import { Provider } from "react-redux";
 import jwt_decode from "jwt-decode";
 
-import store from "./store";
-import { SET_CURRENT_USER } from "./store/actions/types";
-import { getUserContent } from "./store/actions/contentActions";
+import store from "../store";
+import { SET_CURRENT_USER } from "../store/actions/types";
+import { getUserContent } from "../store/actions/contentActions";
 
-import { Home, Login, ProtectedRoute, Register } from "./components";
+import { UserContent, Login, ProtectedRoute, Register } from ".";
 
 if (localStorage.jwt) {
   const jwt = localStorage.jwt;
@@ -29,7 +29,7 @@ function App() {
         <Route exact path="/" component={Login} />
         <Route exact path="/login" component={Login} />
         <Route exact path="/register" component={Register} />
-        <ProtectedRoute exact path="/home" component={Home} />
+        <ProtectedRoute exact path="/home" component={UserContent} />
       </Router>
     </Provider>
   );
