+++ client/src/components/content/index.js
@@ -1,2 +1,2 @@
 export { default as SearchBar } from "./SearchBar";
-export { default as UserContent } from "./UserContent";
+export { default as ContentCard } from "./ContentCard";
